services/servico-agentes-ia/requirements.txt
# Core web framework and server
fastapi
uvicorn
python-multipart

# Asynchronous tasks
celery
redis

# LLM and Vision
llama-cpp-python
instructor
opencv-python-headless
Pillow

# Existing dependencies
confluent-kafka
firebase-admin
python-dotenv